requirements.txt
  # Core
  fastapi==0.115.12
  uvicorn==0.34.1
  python-dotenv==1.1.0
  python-multipart==0.0.20

  # Data
  PyYAML==6.0.2
  openpyxl==3.1.5
  Pillow==10.4.0
  tabulate==0.9.0
  aiofiles==24.1.0

  # Security
  bcrypt==4.3.0
  pyotp==2.9.0
  qrcode==7.4.2
  email_validator==2.2.0

  # Rate limiting
  slowapi==0.1.9
